removed reference only testing classes/layouts.
added format border black to drawables: used as background for view to draw a box around it.

changed servicehost layout to include display views for bike status input from bluetooth (demo working) - not added code yet)

attempted to add underline to strings but it doesnt work in this version of android (no suprise there): given up as taking too long and not important.